// src/app/models/dosya.ts
export class Dosya {
    key: string;
    adi: string;
    uid: string;
    url: string;
    file: File;
    egitimAd: string;
    egitimBilgi: string;
    egitimKategori: string;
    egitimFiyat: string;
    egitimKisaBir: string;
    egitimKisaIki: string;
    egitimKisaUc: string;
    egitimKisaDort: string;
    egitimKisaBes: string;
    egitimKisaAlti: string;
    kayTarih: string;
    duzTarih: string;
}